\name{suda.pal}
\alias{suda.pal}
\title{Color palette generate for suda}
\usage{
  suda.pal(n, type = "seq")
}
\arguments{
  \item{n}{number of colors to return}

  \item{type}{type of color palette (dev, seq, qual)}
}
\description{
  This will return n colors for graphic functions in the
  book.
}
\details{
  Brief description
}
\examples{
color.values <- suda.pal(3, "div")
}
\author{
  Jay Jacobs
}
\keyword{palette}

