\name{suda.pal}
\alias{suda.pal}
\title{suda.pal function}
\usage{
  suda.pal(n, type = "seq")
}
\arguments{
  \item{n.}{number of colors to return}

  \item{type.}{type of color palette}
}
\description{
  function to return colors for the book graphics
}
\details{
  Brief description
}
\examples{
scale_fill_manual(values=suda.pal(3, "div"))
}
\author{
  Jay Jacobs
}
\keyword{palette}

